pyproject.toml
[build-system]
requires = ["flit_core>=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "aoe2_bot"
version = "0.1"
dependencies = [
    "python-telegram-bot>=v20.0",
    "requests>=2.31.0",
]
requires-python = ">=3.8"
authors = [
  { name="Perceval ARENOU", email="perceval.arenou@gmail.com" },
]
description = "A Telegram Bot that gives a random quote of Age of Empires 2 original edition"
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3.8",
    "Topic :: Communications :: Chat",
]

[project.urls]
Homepage = "https://github.com/PercevalSA/TAoEB"
Issues = "https://github.com/PercevalSA/TAoEB/issues"

# [project.scripts]
# aoe-bot = "aoeii_tgbot:main"